Display total spendings of today and this month on dashboard

+++ app/Http/Controllers/TagsController.php
@@ -36,7 +36,10 @@ class TagsController extends Controller {
     }
 
     public function destroy(Tag $tag) {
-        $tag->delete();
+
+        if ($tag->somethign) {
+            $tag->delete();
+        }
 
         return redirect('/settings');
     }
